+++ CHANGES.rst
@@ -16,6 +16,8 @@ APScheduler, see the :doc:`migration section <migration>`.
 
 * Added the ability to modify jobs
 
+* Added the ability to pause and resume jobs
+
 * Dropped the Shelve job store because it could not work with the new job store system
 
 * Adding jobs is now done exclusively through ``add_job()`` -- the shortcuts to triggers were removed